requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bleach==6.0.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
click-log==0.4.0
clickhouse-connect==0.5.22
colorama==0.4.6
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
deprecation==2.1.0
dnspython==2.3.0
docutils==0.19
dotty-dict==1.3.1
duckdb==0.7.1
et-xmlfile==1.1.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.1
Flask-Cors==3.0.10
flask-redis==0.4.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
keyring==23.13.1
langchain==0.0.154
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.1
pipdeptree==2.7.0
pkginfo==1.9.6
postgrest==0.10.6
posthog==3.0.1
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.15.1
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-gitlab==3.14.0
python-magic==0.4.27
python-pptx==0.6.21
python-semantic-release==7.33.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
realtime==1.0.0
redis==4.5.3
regex==2023.3.23
requests==2.29.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
twine==3.8.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.6.2
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.2
win32-setctime==1.1.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
